airbyte-integrations/connectors/source-app-store-connect/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
version = "0.1.0"
name = "source-app-store-connect"
description = "Source implementation for app-store-connect."
authors = ["Airbyte <contact@airbyte.io>"]
license = "MIT"
readme = "README.md"
documentation = "https://docs.airbyte.com/integrations/sources/app-store-connect"
homepage = "https://airbyte.com"
repository = "https://github.com/airbytehq/airbyte"
packages = [{ include = "source_app_store_connect" }, { include = "main.py" }]

[tool.poetry.dependencies]
python = "^3.10,<3.12"
airbyte-cdk = "^4"

[tool.poetry.scripts]
source-app-store-connect = "source_app_store_connect.run:run"

[tool.poetry.group.dev.dependencies]
requests-mock = "*"
pytest-mock = "*"
pytest = "*"
freezegun = "^1.5.1"
